1. Fixed the bug in SubnetTree.addressToInt() 2. Fixed the bug in AddressGenerator.FullRangeGenerator.getNext()

+++ src/main/java/de/rub/nds/tlscrawler/utility/SubnetTree.java
@@ -95,10 +95,10 @@ public class SubnetTree {
 
     private int addressToInt(String ip) {
         String[] address = ip.split(Pattern.quote("."));
-        byte[] ipBytes = new byte[4];
+        int[] ipBytes = new int[4];
 
         for (int i = 0; i < 4; i++) {
-            ipBytes[i] = (byte)Integer.parseInt(address[i]);
+            ipBytes[i] = Integer.parseInt(address[i]);
         }
